requirements.txt
pydantic~=2.11.2
pytest~=8.3.5
httpx~=0.28.1
openai==1.70.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
attrs==25.3.0
black==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
databases==0.9.0
distro==1.9.0
fastapi==0.115.12
flake8==7.2.0
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
isort==6.0.1
jiter==0.9.0
Mako==1.3.9
MarkupSafe==3.0.2
mccabe==0.7.0
multidict==6.3.1
mypy==1.15.0
mypy-extensions==1.0.0
numpy==2.2.4
openai==1.70.0
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
platformdirs==4.3.7
pluggy==1.5.0
propcache==0.3.1
psycopg2==2.9.10
pycodestyle==2.13.0
pydantic_core==2.33.0
pyflakes==3.3.2
pytest==8.3.5
pytest-asyncio==0.26.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.1
structlog==25.2.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.34.0
yarl==1.18.3
